moved flameBearerToTree to separate file
